// Website/Kasino/src/app/interfaces/UserAccount.ts
export interface UserAccount {
  customerID: number | null;
  email: string | null;
  countryID: number | null;
  phoneNumber: number | null;
  firstName: string | null;
  lastName: string | null;
  address: string | null;
  zipCodeID: number | null;
  genderID: number | null;
  registerDate: Date | null;
}